Clean up first line construction logic
Code was duplicated for constructing the first line - an initial block
to attempt it, and then in the main loop if first attempt failed.

Cleaner solution is to have a loop to attempt construction of first line
repeatedly until it succeeds, and then continue into the main loop which
only does expansion/maintenance.