Add handling of old vehicles
Expired vehicles are now handled, albeit somewhat naively. They are sent
to depot and sold when they get there, but not replaced. Since newer
vehicle models may change required # of vehicles on a line, this leaves
it to the next run of DoLineMaintenance to decide if new vehicles are
needed.

Current logic ensures that at least one vehicle is always left, no
matter how old - this works fine for lines that ultimately require more
than one vehicle, but for single-vehicle lines it means it never gets
removed. This will need to be improved.