Honour road vehicle limit
Rather than attempting (and failing) to build new routes or vehicles
when the vehicle limit is reached, check before attempting. In both
cases, this also sets the LastRouteExpansion/LasUpdateDate, so further
attempts are deferred.

Also added a check in Initialize, so the AI will deliberately crash if
there's no road vehicles in the game.