Don't reattempt failed route construction
If building a route fails, don't attempt to use those two towns for the
next attempt.

List of failed towns is cleared after a line is successfully built, so
failures don't permanently make those towns ineligible - this is just a
failsafe to ensure route construction doesn't get stuck in a loop.